CustomList extension
