+++ Shared/Extensions/WatchlistItem-Extension.swift
@@ -166,4 +166,11 @@ extension WatchlistItem: Transferable {
         item.notify = false
         return item
     }
+    
+    var itemLists: Set<CustomList> {
+        let set = list as? Set<CustomList> ?? []
+        return set
+        //return set.sorted(by: {$0.itemTitle < $1.itemTitle})
+    }
 }
+
