+++ Shared/Localization/en.lproj/Localizable.strings
@@ -489,4 +489,5 @@
 "" = "";
 "" = "";
 "" = "";
-
+"addToList" = "Add to List...";
+"addNote" = "Add Note";
