+++ Shared/Localization/pt-BR.lproj/Localizable.strings
@@ -504,3 +504,7 @@
 "createList" = "Criar Nova Lista";
 "yourLists" = "Suas Listas";
 "Lists" = "Listas";
+
+
+"addToList" = "Adicionar à Lista...";
+"addNote" = "Adicionar Nota";
